refactor(lowy): relocate workspace-switcher step to terminal_lifecycle_steps
The 'I select workspace switcher entry {int}' step lived in
theme_steps.ts because of an accidental import overlap. Its volatility
axis is terminal selection — independent of theme — and the new
multi-terminal Code tab scenarios make it a third consumer. Co-locate
it with the existing 'I select terminal {int} in the workspace switcher'
in terminal_lifecycle_steps.ts so the full positional/ID, focus/visibility
matrix of terminal-selection steps lives in one place.
