+++ packages/tests/step_definitions/terminal_lifecycle_steps.ts
@@ -6,8 +6,8 @@ import { Given, Then, When } from "@cucumber/cucumber";
 import { waitForBufferContains } from "../support/buffer.ts";
 import {
   type KoluWorld,
-  WORKSPACE_SWITCHER_ENTRY_SELECTOR,
   POLL_TIMEOUT,
+  WORKSPACE_SWITCHER_ENTRY_SELECTOR,
 } from "../support/world.ts";
 
 When("I create a terminal", async function (this: KoluWorld) {
@@ -49,6 +49,28 @@ When(
       .locator(`[data-terminal-id="${id}"][data-focused]`)
       .waitFor({ state: "attached", timeout: POLL_TIMEOUT });
     await this.waitForFrame();
+  },
+);
+
+/** Select a terminal by its position in the workspace switcher (1-based),
+ *  regardless of `createdTerminalIds`. Complements
+ *  `I select terminal {int} in the workspace switcher` (ID-based, waits on
+ *  `data-focused`): this variant addresses entries by DOM position and
+ *  waits on `data-visible`, which is what scenarios that don't track
+ *  created IDs (or want to address pre-existing background terminals)
+ *  need. */
+When(
+  "I select workspace switcher entry {int}",
+  async function (this: KoluWorld, position: number) {
+    const entry = this.page
+      .locator(WORKSPACE_SWITCHER_ENTRY_SELECTOR)
+      .nth(position - 1);
+    await entry.click();
+    const id = await entry.getAttribute("data-terminal-id");
+    assert.ok(id, `Workspace switcher entry ${position} has no terminal ID`);
+    await this.page
+      .locator(`[data-terminal-id="${id}"][data-visible]`)
+      .waitFor({ state: "attached", timeout: POLL_TIMEOUT });
   },
 );
